[Feat] Food Draw: Basic Gameplay (#9)
* [Feat] Food Draw: Basic Gameplay

The purpose of this MR is to establish the basic functionality for
playing the `Food Draw`. When the page is loaded, 20 locations are
selected. There are 4 rounds. Each round eliminates half the options
until 1 is left.

* [Feat] Defined Win State

The purpose of this MR is to define a `Win State`. Once only 1
restaurant is left, then the win state is trigger. The text at the top
changes to `Winner` and only 1 button is left and it says `Play Again`