[Featt] New Decks and increase randomization
The purpose of this MR is to increase the randomization by creating new
decks. The backend returns 50 entries now. From that 50, half of those
options are selected at random. Thus each time the game is played, a new
deck of options are selected from those 50 items. This process occurs
when the page loads, the game is reset, or a winner has been selected